Normalize short paths to long paths on windows
