+++ src/watchman/watchman.hh
@@ -10,9 +10,7 @@ class WatchmanBackend : public Backend {
 public:
   static bool checkAvailable();
   void start() override;
-  WatchmanBackend() : mStopped(false) {
-    printf("create backend\n");
-  };
+  WatchmanBackend() : mStopped(false) {};
   ~WatchmanBackend();
   void writeSnapshot(Watcher &watcher, std::string *snapshotPath) override;
   void getEventsSince(Watcher &watcher, std::string *snapshotPath) override;
